requirements.txt
amqp==5.2.0
annotated-types==0.7.0
ansicon==1.89.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.10.0
asn1crypto==1.5.1
billiard==4.2.1
blessed==1.20.0
blis==1.3.3
catalogue==2.0.10
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cssselect2==0.7.0
ctranslate2==4.6.2
cymem==2.0.13
distlib==0.3.7
dj-rest-auth==7.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-cron==0.6.0
django-environ==0.11.2
django-filter==25.2
django-picklefield==3.2
django-q==1.3.9
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
emoji==2.15.0
filelock==3.12.3
fsspec==2025.12.0
html5lib==1.1
idna==3.7
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
joblib==1.5.2
kombu==5.4.2
lxml==5.2.1
Markdown==3.9
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
mysqlclient==2.2.7
networkx==3.6.1
numpy==2.3.5
oscrypto==1.3.0
packaging==25.0
pdfkit==1.0.0
pillow==10.3.0
platformdirs==3.10.0
preshed==3.0.12
prompt_toolkit==3.0.48
protobuf==6.33.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
regex==2025.11.3
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.2
sacremoses==0.1.1
sentencepiece==0.2.1
six==1.16.0
smart_open==7.5.0
sqlparse==0.4.4
srsly==2.5.2
stanza==1.10.1
svglib==1.5.1
sympy==1.14.0
thinc==8.3.10
tinycss2==1.3.0
tqdm==4.67.1
typer-slim==0.20.0
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0


virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.3
webencodings==0.5.1
wrapt==2.0.1
xhtml2pdf==0.2.15
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4
